app/requirements.txt
flask
transformers
torch
twilio
python-dotenv